# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Server for WeCare Holidays applications",
        "title": "WeCare Holidays API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.wecareholidays.com/support",
            "email": "support@wecareholidays.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/users": {
            "get": {
                "description": "Get all users with pagination and filtering",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "List users",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "type": "integer",
                        "default": 10,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by full name (partial match, case-insensitive)",
                        "name": "fullName",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by email (partial match, case-insensitive)",
                        "name": "email",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by phone number (partial match, case-insensitive)",
                        "name": "phone",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by status (exact match)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by role ID (exact match)",
                        "name": "roleId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by organization ID (exact match)",
                        "name": "organizationId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search across name, email, and phone",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "default": false,
                        "description": "Include soft-deleted users",
                        "name": "includeDeleted",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.SwaggerStandardResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.PaginatedUsersResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.SwaggerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.SwaggerErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new user with the provided data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Create a new user",
                "parameters": [
                    {
                        "description": "User data",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateUserDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.SwaggerStandardResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/entity.User"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.SwaggerErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/models.SwaggerErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/models.SwaggerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/users/{id}/profile-photo": {
            "post": {
                "description": "Upload a new profile photo for an user and update its record",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Upload user profile photo",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"6824886e6b180b753cea43e9\"",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Logo image file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated user",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.SwaggerStandardResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/entity.User"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/models.SwaggerErrorResponse"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/models.SwaggerErrorResponse"
                        }
                    },
                    "413": {
                        "description": "File too large",
                        "schema": {
                            "$ref": "#/definitions/models.SwaggerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/models.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CreateUserDto": {
            "type": "object",
            "required": [
                "email",
                "fullName",
                "roleId"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "john@example.com"
                },
                "fullName": {
                    "type": "string",
                    "example": "John Doe"
                },
                "organizationId": {
                    "type": "string",
                    "example": "507f1f77bcf86cd799439012"
                },
                "phone": {
                    "type": "string",
                    "example": "+919876543210"
                },
                "profilePhotoUrl": {
                    "type": "string",
                    "example": "https://example.com/photo.jpg"
                },
                "roleId": {
                    "type": "string",
                    "example": "507f1f77bcf86cd799439011"
                },
                "status": {
                    "type": "string",
                    "example": "Active"
                }
            }
        },
...
        "dto.PaginatedUsersResponse": {
            "type": "object",
            "properties": {
                "includeDeleted": {
                    "type": "boolean",
                    "example": false
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.User"
                    }
                },
                "limit": {
                    "type": "integer",
                    "example": 10
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "total": {
                    "type": "integer",
                    "example": 2
                },
                "totalPages": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "entity.AuditTrail": {
            "type": "object",
            "properties": {
                "lastLoginAt": {
                    "type": "string"
                },
                "lastLoginDevice": {
                    "type": "string"
                },
                "lastLoginIp": {
                    "type": "string"
                }
            }
        },
...
        "entity.Email": {
            "type": "object",
            "properties": {
                "email": {
                    "description": "Email address",
                    "type": "string"
                },
                "isVerified": {
                    "description": "Verification status",
                    "type": "boolean"
                }
            }
        },
...
        "entity.Phone": {
            "type": "object",
            "properties": {
                "isVerified": {
                    "description": "Verification status",
                    "type": "boolean"
                },
                "number": {
                    "description": "Phone number (E.164)",
                    "type": "string"
                }
            }
        },
...
        "entity.User": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "auditTrail": {
                    "$ref": "#/definitions/entity.AuditTrail"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "emails": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Email"
                    }
                },
                "fullName": {
                    "type": "string"
                },
                "organizationId": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phones": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Phone"
                    }
                },
                "profilePhotoUrl": {
                    "type": "string"
                },
                "role": {
                    "description": "For convenience",
                    "type": "string"
                },
                "roleId": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/entity.UserStatus"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "entity.UserStatus": {
            "type": "string",
            "enum": [
                "Invited",
                "Active",
                "Suspended",
                "Removed"
            ],
            "x-enum-varnames": [
                "UserStatusInvited",
                "UserStatusActive",
                "UserStatusSuspended",
                "UserStatusRemoved"
            ]
        },
...
        "models.SwaggerErrorResponse": {
            "description": "Error response model",
            "type": "object",
            "properties": {
                "data": {
                    "description": "Optional error details"
                },
                "errorMessage": {
                    "description": "Human-readable error message",
                    "type": "string",
                    "example": "Invalid request parameters"
                },
                "statusCode": {
                    "description": "HTTP status code",
                    "type": "integer",
                    "example": 400
                },
                "success": {
                    "description": "Indicates request failure",
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "models.SwaggerStandardResponse": {
            "description": "Standard API response format",
            "type": "object",
            "properties": {
                "data": {
                    "description": "Response payload"
                },
                "errorMessage": {
                    "description": "Error message (only present if success is false)",
                    "type": "string",
                    "example": ""
                },
                "statusCode": {
                    "description": "HTTP status code",
                    "type": "integer",
                    "example": 200
                },
                "success": {
                    "description": "Indicates whether the request was successful",
                    "type": "boolean",
                    "example": true
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}